Move `URL` extension implementation to its own file
